requirements.txt
lmdb==1.3.0
matplotlib==3.5.2
numpy==1.21.6
opencv_python_headless==4.6.0.66
Pillow==9.2.0
PyYAML==6.0
seaborn==0.11.2
setuptools==59.5.0
tensorboard==2.10.0
tqdm==4.64.0
